+++ pytsite/wallet/_widget.py
@@ -18,14 +18,14 @@ class AccountSelect(_w.select.Select):
 
         if u.has_permission('pytsite.odm_ui.browse.wallet_account') or \
                 u.has_permission('pytsite.odm_ui.browse_own.wallet_account'):
-            f = _odm.find('wallet_account').sort([('aid', _odm.I_ASC)])
+            f = _odm.find('wallet_account').sort([('title', _odm.I_ASC)])
 
             # User can only view its own accounts
             if not u.has_permission('pytsite.odm_ui.browse.wallet_account'):
                 f.where('owner', '=', u)
 
             for acc in f.get():
-                label = '{} ({}, {})'.format(acc.description, acc.aid, acc.currency)
+                label = '{} ({})'.format(acc.title, acc.currency)
                 items.append(('wallet_account:' + str(acc.id), label))
 
         super().__init__(uid, items=items, **kwargs)